requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
gunicorn>=21.2.0
SQLAlchemy-Utils>=0.41.1
Flask-WTF>=1.2
cryptography>=41.0.5
numpy>=1.26
